requirements.txt
# one-piece-scan-bot
pyquery==1.2.13
python-telegram-bot==8.1.1
urlfetch==1.1.3
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth==1.23.0

# heroku-flask
gunicorn==20.0.4
flask==1.1.2
